pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "oisa"
version = "0.1.0"
description = "Orchestrated Immune Simulation Architecture — reference implementation"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.26",
    "scipy>=1.12",
    "pyzmq>=25.1",
    "pyyaml>=6.0",
    "jsonschema>=4.21",
    "networkx>=3.2",
    "pandas>=2.2",
    "matplotlib>=3.8",
    "jupyterlab>=4.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0"]

[tool.hatch.build.targets.wheel]
packages = ["analysis", "orchestrator", "models"]

[tool.pytest.ini_options]
testpaths = ["models", "orchestrator"]
python_files = ["test_*.py"]
